markdown-converter/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9
pydantic>=2.9.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
google-cloud-storage>=2.18.0
google-cloud-aiplatform>=1.71.0
google-auth>=2.36.0
itsdangerous>=2.2.0
mistralai[gcp]>=1.2.0
pypdf>=5.0.0
google-api-python-client>=2.100.0
